Patch 5.22 mastery display changes.
Changes how masteries are displayed for preseason masteries in Patch 5.22 so that every odd numbered row will have an extra spacer between the masteries and 2 will always be displayed on every odd row. Checking for the count will no longer be necessary for any row because alignment will also be changed to centered, so it is not needed.